requirements.txt
torch
transformers
huggingface_hub
opencv-python
pyautogui
psutil
numpy
pillow
pyvips
einops
pygetwindow
pywin32
timm
accelerate>=0.26.0
flash-attn
pydirectinput
pyvjoy
mss
